Extract registrations link logic
