+++ domains/eventEditor/src/ui/tickets/TicketRegistrationsLink.tsx
@@ -1,12 +1,9 @@
-import React, { useMemo } from 'react';
-import { addQueryArgs } from '@wordpress/url';
+import React from 'react';
 import { __ } from '@eventespresso/i18n';
 
-import { ADMIN_ROUTES } from '@eventespresso/constants';
-import { getAdminUrl, useEventId } from '@eventespresso/edtr-services';
-import { useConfig } from '@eventespresso/services';
 import { RegistrationsLink, ItemCount } from '@eventespresso/components';
-import { Ticket } from '@eventespresso/edtr-services';
+import type { Ticket } from '@eventespresso/edtr-services';
+import { useRegistrationsLink } from '@eventespresso/edtr-services';
 
 interface Props {
 	ticket: Ticket;
@@ -15,21 +12,7 @@ interface Props {
 const tooltipProps = { placement: 'top' as const };
 
 const TicketRegistrationsLink: React.FC<Props> = ({ ticket }) => {
-	const { siteUrl } = useConfig();
-
-	const eventId = useEventId();
-
-	const regListUrl = useMemo(() => {
-		const adminUrl = getAdminUrl({
-			adminSiteUrl: siteUrl.admin,
-			page: ADMIN_ROUTES.REGISTRATIONS,
-		});
-		return addQueryArgs(adminUrl, {
-			event_id: eventId,
-			ticket_id: ticket.dbId,
-			return: 'edit',
-		});
-	}, [eventId, siteUrl.admin, ticket.dbId]);
+	const regListUrl = useRegistrationsLink({ ticket_id: ticket.dbId });
 
 	const countTitle = __('total registrations.');
 	const tooltip = __('view ALL registrations for this ticket.');
