+++ packages/edtr-services/src/hooks/index.ts
@@ -1,2 +1,3 @@
 export * from './edtrState';
 export * from './entityListCacheUpdates';
+export * from './useRegistrationsLink';
